refactor(modules): extract the git/forge subsystem into src/modules/git/
Consolidate the whole git/forge surface into one module, mirroring the
economizer/protocols/vault relocations, and absorbing the mcp_git_* tools
deferred from the protocols move (#1514):

modules/git/ <-
  server/git_*.c (13: cred_inject, forge_vault, host_cred, host_resolve,
    oauth_{device,gh,github}, ops, org_repos, pr_api, pr_ci_grade, project,
    ssh_agent)
  git_verify*.c (8: the git-operation verification subsystem)
  forge_credentials.c
  mcp_git_{branch,pr,query,write}.c (the MCP git tools)
  + their 18 headers.

Pure relocate: git mv + build-path updates + include rewrites, no logic
change. The 12 top-level moved files used `#include "headers/*"` prefixes,
rewritten to bare (resolve via the new -Imodules/git); the 13 server/git_*
files already used bare includes. 8 stay-behind includers (cmd_*,
guardrails*, agent_pipeline, wfe_live_delegate) had headers/-prefixed
includes of the moved git_verify.h / mcp_git.h, rewritten to bare. Both
build systems + the MCP_GIT_SRCS bundle updated (Makefile C_FLAGS,
MCP_GIT_SRCS, CORE_SRCS, SERVER_SRCS, BENCH_OBJS; tests/Rules.mk object
paths; CMake sources + all six target_include_directories) and the
check-git-cred-centralized.py exempt-path allow-list.

Stays behind (different concern): server/server_http_routes_git.c (HTTP
routes), server/wfe_live_forge.c, forge_app_token.c (function-pointer seam),
db1/git_ownership.c, and the tests/support git stubs.

Server links; unit-test-git-{cred-inject,forge-vault,host-resolve,ops,
pr-ci-grade,project,ssh-agent,verify-contract,verify-select} + mcp-git +
forge-credentials pass; make build-integrity + make lint green.
