+++ src/cmd_core.c
@@ -17,8 +17,8 @@
 #include "memory.h"
 #include "cJSON.h"
 #include "aimee_client.h"
-#include "headers/mcp_git.h"
-#include "headers/git_verify.h"
+#include "mcp_git.h"
+#include "git_verify.h"
 #include "headers/events.h"
 #include "headers/aimee_home.h"
 #include "headers/workspace_manifest.h"
