+++ src/guardrails_orchestrator.c
@@ -15,7 +15,7 @@
 #include "guardrails_blast_radius.h"
 #include "workspace_provider.h" /* skip worktree enforcement for a detached (client) workspace */
 #include "headers/config.h"
-#include "headers/git_verify.h"
+#include "git_verify.h"
 #include "headers/skill.h"
 #include "kb_client.h"
 #if !defined(AIMEE_DB2_DISABLED)
