+++ src/modules/git/git_cred_inject.c
