+++ src/modules/git/git_host_cred.c
