+++ src/modules/git/git_host_resolve.c
