+++ src/modules/git/git_host_resolve.h
