+++ src/modules/git/git_oauth_device.h
