+++ src/modules/git/git_verify.h
