+++ src/modules/git/git_verify_internal.h
