+++ src/modules/git/git_verify_jobs.h
