+++ src/modules/git/git_verify_step.c
@@ -1,8 +1,8 @@
-#include "headers/git_verify_internal.h"
+#include "git_verify_internal.h"
 
-#include "headers/dstr.h"
-#include "headers/platform_process.h"
-#include "headers/util.h"
+#include "dstr.h"
+#include "platform_process.h"
+#include "util.h"
 
 #include <stdio.h>
 #include <stdlib.h>
