+++ src/modules/git/mcp_git.h
