+++ src/modules/git/mcp_git_query.c
@@ -2,15 +2,15 @@
 #include "aimee.h"
 #include "db1/git_ownership.h"
 #include "cJSON.h"
-#include "headers/config.h"
-#include "headers/guardrails.h"
-#include "headers/git_verify.h"
-#include "headers/mcp_git.h"
-#include "headers/platform_process.h"
-#include "headers/util.h"
-#include "headers/workspace_provider.h"
-#include "headers/forge_credentials.h"
-#include "headers/git_cred_inject.h"
+#include "config.h"
+#include "guardrails.h"
+#include "git_verify.h"
+#include "mcp_git.h"
+#include "platform_process.h"
+#include "util.h"
+#include "workspace_provider.h"
+#include "forge_credentials.h"
+#include "git_cred_inject.h"
 #include <time.h>
 
 extern char **environ;
